requirements.txt
arrow==1.3.0
asgiref==3.8.1
binaryornot==0.4.4
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
Django==5.1.1
django-admin-interface==0.28.9
django-browser-reload==1.15.0
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-js-asset==2.2.0
django-tailwind==3.8.0
idna==3.10
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
pillow==10.4.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.2
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20241003
urllib3==2.2.3
